backend/requirements.txt
django==4.1.2
djangorestframework
django-cors-headers
django-cockroachdb==4.1
gunicorn
whitenoise
dj-database-url
psycopg2-binary==2.9.4